Forwarder form ported - the simplest one (refs #20)
